+++ Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/filipdulic/bus-queue"
 readme = "README.md"
 
 [dependencies]
-lockfree = "^0.5.1"
+lockfree = {version = "^0.5.1", optional = true}
 arc-swap = "0.3.6"
 futures = {version = "0.1", optional = true}
 tokio = {version = "0.1", optional = true}
@@ -17,7 +17,9 @@ tokio = {version = "0.1", optional = true}
 [features]
 default = ["async"]
 async = ["futures"]
+sync = ["lockfree"]
 async-example = ["tokio"]
+
 
 [[example]]
 name = "bare-simple"
