+++ src/async_.rs
@@ -1,6 +1,8 @@
 use super::*;
 use futures::prelude::*;
 use futures::{task::AtomicTask, Async, AsyncSink};
+use futures::sync::mpsc;
+use futures::task;
 
 #[derive(Debug)]
 pub struct Publisher<T: Send> {
@@ -19,6 +21,7 @@ pub fn channel<T: Send>(size: usize) -> (Publisher<T>, Subscriber<T>) {
     (
         Publisher {
             bare_publisher,
+            task: task::current(),
             waker,
         },
         Subscriber {
@@ -56,14 +59,13 @@ impl<T: Send> Sink for Publisher<T> {
         Ok(Async::Ready(()))
     }
     fn close(&mut self) -> Poll<(), Self::SinkError> {
-        Ok(Async::Ready(()))
+        self.poll_complete()
     }
 }
 
 impl<T: Send> Drop for Publisher<T> {
     fn drop(&mut self) {
         self.close().unwrap();
-        self.poll_complete().unwrap();
     }
 }
 
@@ -78,6 +80,7 @@ impl<T: Send> Eq for Publisher<T> {}
 impl<T: Send> Stream for Subscriber<T> {
     type Item = Arc<T>;
     type Error = ();
+
     fn poll(&mut self) -> Poll<Option<Self::Item>, Self::Error> {
         match self.bare_subscriber.try_recv() {
             Ok(arc_object) => Ok(Async::Ready(Some(arc_object))),
@@ -95,7 +98,7 @@ impl<T: Send> Stream for Subscriber<T> {
 impl<T: Send> Clone for Subscriber<T> {
     fn clone(&self) -> Self {
         let arc_t = Arc::new(AtomicTask::new());
-        self.sleeper.sender.send(arc_t.clone()).unwrap();
+        self.sleeper.sender.send(arc_t.clone());
         Self {
             bare_subscriber: self.bare_subscriber.clone(),
             sleeper: Sleeper {
@@ -113,3 +116,49 @@ impl<T: Send> PartialEq for Subscriber<T> {
 }
 
 impl<T: Send> Eq for Subscriber<T> {}
+
+/// Helper struct used by sync and async implementations to wake Tasks / Threads
+#[derive(Debug)]
+pub struct Waker<T> {
+    /// Vector of Tasks / Threads to be woken up.
+    pub sleepers: Vec<Arc<T>>,
+    /// A mpsc Receiver used to receive Tasks / Threads to be registered.
+    receiver: mpsc::Receiver<Arc<T>>,
+}
+
+/// Helper struct used by sync and async implementations to register Tasks / Threads to
+/// be woken up.
+#[derive(Debug)]
+pub struct Sleeper<T> {
+    /// Current Task / Thread to be woken up.
+    pub sleeper: Arc<T>,
+    /// mpsc Sender used to register Task / Thread.
+    pub sender: mpsc::Sender<Arc<T>>,
+}
+
+impl<T> Waker<T> {
+    /// Register all the Tasks / Threads sent for registration.
+    pub fn register_receivers(&mut self) -> impl Future<Item=()> {
+        for receiver in self.receiver.recv() {
+            self.sleepers.push(receiver);
+        }
+    }
+}
+
+/// Function used to create a ( Waker, Sleeper ) tuple.
+pub fn alarm<T>(current: T) -> (Waker<T>, Sleeper<T>) {
+    let mut vec = Vec::new();
+    let (sender, receiver) = mpsc::channel(10);
+    let arc_t = Arc::new(current);
+    vec.push(arc_t.clone());
+    (
+        Waker {
+            sleepers: vec,
+            receiver,
+        },
+        Sleeper {
+            sleeper: arc_t,
+            sender,
+        },
+    )
+}
