+++ src/sync.rs
@@ -152,3 +152,58 @@ impl<T: Send> PartialEq for Subscriber<T> {
 }
 
 impl<T: Send> Eq for Subscriber<T> {}
+
+/// Helper struct used by sync and async implementations to wake Tasks / Threads
+#[derive(Debug)]
+pub struct Waker<T> {
+    /// Vector of Tasks / Threads to be woken up.
+    pub sleepers: Vec<Arc<T>>,
+    /// A mpsc Receiver used to receive Tasks / Threads to be registered.
+    receiver: lockfree::channel::mpsc::Receiver<Arc<T>>,
+}
+
+/// Helper struct used by sync and async implementations to register Tasks / Threads to
+/// be woken up.
+#[derive(Debug)]
+pub struct Sleeper<T> {
+    /// Current Task / Thread to be woken up.
+    pub sleeper: Arc<T>,
+    /// mpsc Sender used to register Task / Thread.
+    pub sender: lockfree::channel::mpsc::Sender<Arc<T>>,
+}
+
+impl<T> Waker<T> {
+    /// Register all the Tasks / Threads sent for registration.
+    pub fn register_receivers(&mut self) {
+        for receiver in self.receiver.recv() {
+            self.sleepers.push(receiver);
+        }
+    }
+}
+
+/// Function used to create a ( Waker, Sleeper ) tuple.
+pub fn alarm<T>(current: T) -> (Waker<T>, Sleeper<T>) {
+    let mut vec = Vec::new();
+    let (sender, receiver) = lockfree::channel::mpsc::create();
+    let arc_t = Arc::new(current);
+    vec.push(arc_t.clone());
+    (
+        Waker {
+            sleepers: vec,
+            receiver,
+        },
+        Sleeper {
+            sleeper: arc_t,
+            sender,
+        },
+    )
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    
+    #[test]
+    fn channel() {
+    }
+}
